requirements.txt
tensorflow==2.20.0
numpy
pillow
streamlit
gdown